Fix call to overloded property

+++ src/algorithms/signal_source/adapters/file_signal_source.cc
@@ -41,8 +41,8 @@ FileSignalSource::FileSignalSource(const ConfigurationInterface* configuration,
 
     double default_seconds_to_skip = 0.0;
     size_t header_size = 0;
-    samples_ = configuration->property(role + ".samples", 0ULL);
-    sampling_frequency_ = configuration->property(role + ".sampling_frequency", 0LL);
+    samples_ = configuration->property(role + ".samples", static_cast<uint64_t>(0));
+    sampling_frequency_ = configuration->property(role + ".sampling_frequency", static_cast<int64_t>(0));
     filename_ = configuration->property(role + ".filename", default_filename);
 
     // override value with commandline flag, if present
