+++ src/algorithms/signal_source/adapters/osmosdr_signal_source.cc
@@ -37,7 +37,7 @@ OsmosdrSignalSource::OsmosdrSignalSource(const ConfigurationInterface* configura
     std::string empty = "";
     std::string default_dump_file = "./data/signal_source.dat";
     std::string default_item_type = "gr_complex";
-    samples_ = configuration->property(role + ".samples", 0LL);
+    samples_ = configuration->property(role + ".samples", static_cast<int64_t>(0));
     dump_ = configuration->property(role + ".dump", false);
     dump_filename_ = configuration->property(role + ".dump_filename",
         default_dump_file);
