+++ src/algorithms/signal_source/adapters/spir_gss6450_file_signal_source.cc
@@ -35,8 +35,8 @@ SpirGSS6450FileSignalSource::SpirGSS6450FileSignalSource(const ConfigurationInte
     std::string default_dump_filename = "../data/my_capture_dump.dat";
     item_type_ = "int";
 
-    samples_ = configuration->property(role + ".samples", 0ULL);
-    sampling_frequency_ = configuration->property(role + ".sampling_frequency", 0LL);
+    samples_ = configuration->property(role + ".samples", static_cast<uint64_t>(0));
+    sampling_frequency_ = configuration->property(role + ".sampling_frequency", static_cast<int64_t>(0));
     filename_ = configuration->property(role + ".filename", default_filename);
     repeat_ = configuration->property(role + ".repeat", false);
     dump_ = configuration->property(role + ".dump", false);
